[FEATURE] Make creating additional doctrine connections easier
This is mainly a cleanup of the ``EntityManagerFactory`` and the
``FlowAnnotationDriver`` splitting both up into a base version to
be used as basis for custom doctrine connections and a Flow specific
version with the same class name as before. Additionally a lot of the
code was refactored into smaller methods to improve maintainability
and readability. No functional changes have been done to the code.

Entities using the ``BaseAnnotationDriver`` would not get the
``Persistent_Object_Identifier`` introduced which makes sense
for non Flow entities.

Resolves: FLOW-391
