+++ TYPO3.Flow/Classes/TYPO3/Flow/Persistence/Doctrine/EntityManagerFactory.php
@@ -8,41 +8,29 @@ namespace TYPO3\Flow\Persistence\Doctrine;
  * the terms of the MIT license.                                          *
  *                                                                        */
 
-use Doctrine\DBAL\Types\Type;
 use Doctrine\ORM\Configuration;
 use TYPO3\Flow\Annotations as Flow;
 use TYPO3\Flow\Configuration\Exception\InvalidConfigurationException;
-use TYPO3\Flow\Persistence\Exception\IllegalObjectTypeException;
+use TYPO3\Flow\Persistence\Doctrine\Mapping\Driver\FlowAnnotationDriver;
+use TYPO3\Flow\Utility\Files;
 
 /**
  * EntityManager factory for Doctrine integration
  *
  * @Flow\Scope("singleton")
  */
-class EntityManagerFactory
+class EntityManagerFactory extends BaseEntityManagerFactory
 {
     /**
-     * @Flow\Inject
-     * @var \TYPO3\Flow\Object\ObjectManagerInterface
+     * @var array
      */
-    protected $objectManager;
-
-    /**
-     * @Flow\Inject
-     * @var \TYPO3\Flow\Reflection\ReflectionService
-     */
-    protected $reflectionService;
+    protected $settings = [];
 
     /**
      * @Flow\Inject
      * @var \TYPO3\Flow\Utility\Environment
      */
     protected $environment;
-
-    /**
-     * @var array
-     */
-    protected $settings = array();
 
     /**
      * Injects the Flow settings, the persistence part is kept
@@ -54,140 +42,48 @@ class EntityManagerFactory
      */
     public function injectSettings(array $settings)
     {
+        if (!isset($settings['persistence'])) {
+            throw new InvalidConfigurationException('The configuration TYPO3.Flow.persistence is NULL, please check your settings.', 1392800005);
+        }
         $this->settings = $settings['persistence'];
-        if (!is_array($this->settings['doctrine'])) {
-            throw new InvalidConfigurationException(sprintf('The TYPO3.Flow.persistence.doctrine settings need to be an array, %s given.', gettype($this->settings['doctrine'])), 1392800005);
-        }
-        if (!is_array($this->settings['backendOptions'])) {
-            throw new InvalidConfigurationException(sprintf('The TYPO3.Flow.persistence.backendOptions settings need to be an array, %s given.', gettype($this->settings['backendOptions'])), 1426149224);
-        }
     }
 
     /**
      * Factory method which creates an EntityManager.
      *
+     * @param array $settings
+
      * @return \Doctrine\ORM\EntityManager
-     * @throws \TYPO3\Flow\Configuration\Exception\InvalidConfigurationException
+     * @see BaseEntityManagerFactory::create
      */
-    public function create()
+    public function create(array $settings = [])
     {
-        $config = new Configuration();
-        $config->setClassMetadataFactoryName(\TYPO3\Flow\Persistence\Doctrine\Mapping\ClassMetadataFactory::class);
-
-        $cache = new \TYPO3\Flow\Persistence\Doctrine\CacheAdapter();
-        // must use ObjectManager in compile phase...
-        $cache->setCache($this->objectManager->get(\TYPO3\Flow\Cache\CacheManager::class)->getCache('Flow_Persistence_Doctrine'));
-        $config->setMetadataCacheImpl($cache);
-        $config->setQueryCacheImpl($cache);
-
-        $resultCache = new \TYPO3\Flow\Persistence\Doctrine\CacheAdapter();
-        // must use ObjectManager in compile phase...
-        $resultCache->setCache($this->objectManager->get(\TYPO3\Flow\Cache\CacheManager::class)->getCache('Flow_Persistence_Doctrine_Results'));
-        $config->setResultCacheImpl($resultCache);
-
-        if (is_string($this->settings['doctrine']['sqlLogger']) && class_exists($this->settings['doctrine']['sqlLogger'])) {
-            $sqlLoggerInstance = new $this->settings['doctrine']['sqlLogger']();
-            if ($sqlLoggerInstance instanceof \Doctrine\DBAL\Logging\SQLLogger) {
-                $config->setSQLLogger($sqlLoggerInstance);
-            } else {
-                throw new InvalidConfigurationException(sprintf('TYPO3.Flow.persistence.doctrine.sqlLogger must point to a \Doctrine\DBAL\Logging\SQLLogger implementation, %s given.', get_class($sqlLoggerInstance)), 1426150388);
-            }
-        }
-
-        $eventManager = $this->buildEventManager();
-
-        $flowAnnotationDriver = $this->objectManager->get(\TYPO3\Flow\Persistence\Doctrine\Mapping\Driver\FlowAnnotationDriver::class);
-        $config->setMetadataDriverImpl($flowAnnotationDriver);
-
-        $proxyDirectory = \TYPO3\Flow\Utility\Files::concatenatePaths(array($this->environment->getPathToTemporaryDirectory(), 'Doctrine/Proxies'));
-        \TYPO3\Flow\Utility\Files::createDirectoryRecursively($proxyDirectory);
-        $config->setProxyDir($proxyDirectory);
-        $config->setProxyNamespace('TYPO3\Flow\Persistence\Doctrine\Proxies');
-        $config->setAutoGenerateProxyClasses(false);
-
-        // The following code tries to connect first, if that succeeds, all is well. If not, the platform is fetched directly from the
-        // driver - without version checks to the database server (to which no connection can be made) - and is added to the config
-        // which is then used to create a new connection. This connection will then return the platform directly, without trying to
-        // detect the version it runs on, which fails if no connection can be made. But the platform is used even if no connection can
-        // be made, which was no problem with Doctrine DBAL 2.3. And then came version-aware drivers and platforms...
-        $connection = \Doctrine\DBAL\DriverManager::getConnection($this->settings['backendOptions'], $config, $eventManager);
-        try {
-            $connection->connect();
-        } catch (\Doctrine\DBAL\Exception\ConnectionException $exception) {
-            $settings = $this->settings['backendOptions'];
-            $settings['platform'] = $connection->getDriver()->getDatabasePlatform();
-            $connection = \Doctrine\DBAL\DriverManager::getConnection($settings, $config, $eventManager);
-        }
-
-        $entityManager = \Doctrine\ORM\EntityManager::create($connection, $config, $eventManager);
-        $flowAnnotationDriver->setEntityManager($entityManager);
-
-        if (isset($this->settings['doctrine']['dbal']['mappingTypes']) && is_array($this->settings['doctrine']['dbal']['mappingTypes'])) {
-            foreach ($this->settings['doctrine']['dbal']['mappingTypes'] as $typeName => $typeConfiguration) {
-                Type::addType($typeName, $typeConfiguration['className']);
-                $entityManager->getConnection()->getDatabasePlatform()->registerDoctrineTypeMapping($typeConfiguration['dbType'], $typeName);
-            }
-        }
-
-        if (isset($this->settings['doctrine']['filters']) && is_array($this->settings['doctrine']['filters'])) {
-            foreach ($this->settings['doctrine']['filters'] as $filterName => $filterClass) {
-                $config->addFilter($filterName, $filterClass);
-                $entityManager->getFilters()->enable($filterName);
-            }
-        }
-
-        if (isset($this->settings['doctrine']['dql']) && is_array($this->settings['doctrine']['dql'])) {
-            $this->applyDqlSettingsToConfiguration($this->settings['doctrine']['dql'], $config);
-        }
-
-        return $entityManager;
+        $settings = array_merge($this->settings, $settings);
+        return parent::create($settings);
     }
 
     /**
-     * Add configured event subscribers and listeners to the event manager
+     * Get the annotation driver responsible for this connection.
      *
-     * @return \Doctrine\Common\EventManager
-     * @throws \TYPO3\Flow\Persistence\Exception\IllegalObjectTypeException
+     * @param array $settings
+     * @return FlowAnnotationDriver
      */
-    protected function buildEventManager()
+    protected function getAnnotationDriver(array $settings = [])
     {
-        $eventManager = new \Doctrine\Common\EventManager();
-        if (isset($this->settings['doctrine']['eventSubscribers']) && is_array($this->settings['doctrine']['eventSubscribers'])) {
-            foreach ($this->settings['doctrine']['eventSubscribers'] as $subscriberClassName) {
-                $subscriber = $this->objectManager->get($subscriberClassName);
-                if (!$subscriber instanceof \Doctrine\Common\EventSubscriber) {
-                    throw new IllegalObjectTypeException('Doctrine eventSubscribers must extend class \Doctrine\Common\EventSubscriber, ' . $subscriberClassName . ' fails to do so.', 1366018193);
-                }
-                $eventManager->addEventSubscriber($subscriber);
-            }
-        }
-        if (isset($this->settings['doctrine']['eventListeners']) && is_array($this->settings['doctrine']['eventListeners'])) {
-            foreach ($this->settings['doctrine']['eventListeners'] as $listenerOptions) {
-                $listener = $this->objectManager->get($listenerOptions['listener']);
-                $eventManager->addEventListener($listenerOptions['events'], $listener);
-            }
-        }
-        return $eventManager;
+        return $this->objectManager->get(FlowAnnotationDriver::class);
     }
 
     /**
-     * Apply configured settings regarding DQL to the Doctrine Configuration.
-     * At the moment, these are custom DQL functions.
-     *
-     * @param array $configuredSettings
      * @param Configuration $doctrineConfiguration
+     * @param array $settings
      * @return void
      */
-    protected function applyDqlSettingsToConfiguration(array $configuredSettings, Configuration $doctrineConfiguration)
+    protected function setProxyPath(Configuration $doctrineConfiguration, $settings)
     {
-        if (isset($configuredSettings['customStringFunctions'])) {
-            $doctrineConfiguration->setCustomStringFunctions($configuredSettings['customStringFunctions']);
+        if (!isset($settings['doctrine']['proxyPath'])) {
+            $settings['doctrine']['proxyPath'] = Files::concatenatePaths([$this->environment->getPathToTemporaryDirectory(), 'Doctrine/Proxies']);
         }
-        if (isset($configuredSettings['customNumericFunctions'])) {
-            $doctrineConfiguration->setCustomNumericFunctions($configuredSettings['customNumericFunctions']);
-        }
-        if (isset($configuredSettings['customDatetimeFunctions'])) {
-            $doctrineConfiguration->setCustomDatetimeFunctions($configuredSettings['customDatetimeFunctions']);
-        }
+
+        parent::setProxyPath($doctrineConfiguration, $settings);
     }
 }
